Attempt to resolve lighterhtml+dom-augmentor's issues by updating libs
The lighterhtml+dom-augmentor libraries re-use instances of component heavily,
even when not desired. In this case the drag&drop does not work for the second
or any subsequent drawn card because the context update triggers an update for
the previous instance of the draggable HOC instead of the currently rendered.

This commit is used to mark the moment at which I gave up on using lighterhtml
and dom-augmentor directly. This marks the start of the attempt at using
neverland which should handle this integration for us and - hopefully - work
better.